* typhon/arts/catalogues.py: Reimplemented Sparse class.
The Sparse class is now a simple wrapper around SciPy's csc_matrix.

The usage of both classes is exactly the same, the only differences are
additional reading and writing functions for XML files as well as
ARTS-consistent attributenames.

Changes are **not** backwards compatible! Code relying on the old
implementation of Sparse will fail and has to be adapted.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/rt/typhon/trunk@9989 aaf1aab0-4228-0410-ad68-8dceda47f409