* doc/pyplots/plot_difference.py:
Added example plot for the "difference" colormap.

* doc/pyplots/plot_phase.py:

Improved the overall appearance of the plot by adding a colorbar
and decreasing the number of quivers to draw.

* doc/pyplots/plot_qualitative1.py:

Set xlimits to improve the look of the example plot.

* doc/typhon.plots.cm.rst:

Added the example plot for the "difference" colormap.

Alphabetized the order of the example plots.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/rt/typhon/trunk@10384 aaf1aab0-4228-0410-ad68-8dceda47f409