* typhon/physics/em.py (planck, planck_wavelength, planck_wavenumber, rayleighjeans_wavelength, radiance2planckTb, radiance2rayleighjeansTb, frequency2wavelength, wavelength2frequency, wavelength2wavenumber, wavenumber2wavelength):
Introduce a np.divide to avoid a runtime error being thrown if function
is called with a float instead of an numpy array. Now for all such
cases a numpy warning is consistently issued.

This makes the functions behave in line with how numpy's / scipy's own
functions work. If the user wants to suppress the warning, the
means numpy offers itself should be used for this, e.g.:

with np.errstate(divide='ignore'):
    radiance2plancktb(f, r)

git-svn-id: https://arts.mi.uni-hamburg.de/svn/rt/typhon/trunk@10498 aaf1aab0-4228-0410-ad68-8dceda47f409