Improved usabilty of Timer class.
* typhon/utils/__init__.py (Timer):

It is possible to pass additional information on the process
being timed using the `info` keyword. If no information is
given, a default value depending on the context is used e.g. the
function name when used as decorator.

It is also possible to alter the timeformat.

Removed unnecessary lines and updated docstrings.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/rt/typhon/trunk@10590 aaf1aab0-4228-0410-ad68-8dceda47f409