Reimplement handling of environment variables.
* typhon/environment.py:

Reimplement the whole handling of environment variables.

Remove all module-level constants, as the fact that they are constant is
limiting the use case in scripts.

The new implementation is based on the behvaiour os `os.environ`.

* typhon/config.py: Simplify the whole implementation.

Make the config case-sensitive. This is needed to be consistent with
environment variables.

* typhon/__init__.py: Import `typhon.environ` on top-level.

* typhon/arts/common.py, typhon/arts/workspace/__init__.py,
  typhon/arts/workspace/api.py:

Adapt to the new API of `typhon.environ`.

* typhon/tests/test_environment.py: Add tests for membership check.

* doc/conf.py, doc/typhon.config.rst, doc/typhon.environment.rst:

Update documentation to new module structure.