Function enabling DataArray.stack in case of repeated dimensions
When a dimension is repeated in a DataArray (such as may be the case
when storing covariance matrices), several of xarrays methods do not
currently work.  This commit adds to typhon a workaround to enable
DataArray.stack to work for DataArray objects where a dimension occurs
multiple times.  Use with care, still early days.