Add necessary version for xarray dependency
Since commit c6ea042 (merged in #143), UnitsAwareDataArray depends on
xarray.DataArray.__array_ufunc__ (which in turn depends on numpy 1.13 or
newer).  This was merged into xarray in
https://github.com/pydata/xarray/pull/1962 and added to release 0.10.2.
The old UnitsAwareDataArray does not work with xarray>=0.10.2 and the new
one does not work with xarray<=0.10.1.  Therefore, typhon now depends on
xarray>=0.10.2 and numpy>=1.13.