Made SRF.estimate_band_coefficients even more fragile
SRF.estimate_band_coefficients should not be touched even with a long
stick.  It reads band coefficients from a set of fragilly defined files
with fragilly defined names.  This is the precariat pur sang if there
ever was one.  It is a house of cards that is about to collapse and should
probably be nuked from orbit, yet instead I have had the audacity to add
yet another layer of fragility to it.  For love and peace it is probably
best to pretend this commit does not exist and that the method I have
changed in the commit has never existed either.