Make adaptations to actually support other filesystems
Make necessary adaptations to actually support other file systems.
Not all file systems have a root (s3fs, zipfile do not), and if they
don't, then the FileHandler insistance on absolute paths means it's not
possible for any file to exist on such a system.  The tests pass on my
system on Python 3.8.