Loosen error margin for Timer test
It happened that the Github CI failed on Windows because the Timer was
shorter than the time the test should have slept. This is likely caused
by inaccuracies in the `sleep()` implementation.  As a workaround set a
specific error margin which also allows the measured time to be slightly
smaller than the given sleep time.